+++ backend/app/__init__.py
@@ -79,8 +79,8 @@ def _load_data(data_loader):
     data_loader.load_contact_questionnaire()
     data_loader.load_demographics_questionnaire()
     data_loader.load_evaluation_history_questionnaire()
-    data_loader.load_guardian_demographics_questionnaire()
     data_loader.load_home_questionnaire()
+    data_loader.load_identification_questionnaire()
 
 
 @app.cli.command()
