+++ backend/app/data_loader.py
@@ -14,7 +14,7 @@ from app.model.questionnaires.contact_questionnaire import ContactQuestionnaire
 from app.model.questionnaires.demographics_questionnaire import DemographicsQuestionnaire
 from app.model.questionnaires.evaluation_history_questionnaire import EvaluationHistoryQuestionnaire
 from app.model.questionnaires.home_questionnaire import HomeQuestionnaire
-from app.model.questionnaires.guardian_demographics_questionnaire import GuardianDemographicsQuestionnaire
+from app.model.questionnaires.identification_questionnaire import IdentificationQuestionnaire
 from app import db
 from sqlalchemy import Sequence
 import csv
@@ -136,17 +136,15 @@ class DataLoader():
         db.session.commit()
 
     def load_contact_questionnaire(self):
-        c_ques = ContactQuestionnaire(first_name="Charlie", last_name="Brown")
+        c_ques = ContactQuestionnaire(phone=555-555-1234, contact_times='Weekdays at 5AM', email='charlie@brown.com')
         db.session.add(c_ques)
         print("Contact loaded.  There is now %i contact record in the database." % db.session.query(
             ContactQuestionnaire).count())
         db.session.commit()
 
     def load_demographics_questionnaire(self):
-        d_ques = DemographicsQuestionnaire(guardian_id=1, participant_id=2, first_name="Charles", middle_name="Monroe",
-                                         last_name="Brown", is_first_name_preferred=False, nickname="Charlie",
-                                         birthdate="1979-1-5", birth_city="Staunton", birth_state="VA", birth_sex='male',
-                                         gender_identity='male', race_ethnicity="raceWhite", is_english_primary=True)
+        d_ques = DemographicsQuestionnaire(user_id=1, birth_sex='male', gender_identity='intersex',
+                                           race_ethnicity="raceAsian")
         db.session.add(d_ques)
         print("Demographics loaded.  There is now %i demographics record in the database." % db.session.query(
             DemographicsQuestionnaire).count())
@@ -166,11 +164,13 @@ class DataLoader():
             HomeQuestionnaire).count())
         db.session.commit()
 
-    def load_guardian_demographics_questionnaire(self):
-        gd_ques = GuardianDemographicsQuestionnaire(birthdate="1979-1-5", sex='male', race_ethnicity="raceWhite", is_english_primary=True)
-        db.session.add(gd_ques)
-        print("Guardian Demographics loaded.  There is now %i guardian demographics record in the database." % db.session.query(
-            GuardianDemographicsQuestionnaire).count())
+    def load_identification_questionnaire(self):
+        i_ques = IdentificationQuestionnaire(first_name="Charles", middle_name="Monroe", last_name="Brown",
+                                             is_first_name_preferred=False, nickname="Charlie", birthdate="1979-1-5",
+                                             birth_city="Staunton", birth_state="VA", is_english_primary=True)
+        db.session.add(i_ques)
+        print("Identification loaded.  There is now %i home record in the database." % db.session.query(
+            IdentificationQuestionnaire).count())
         db.session.commit()
 
     def get_org_by_name(self, org_name):
@@ -193,8 +193,8 @@ class DataLoader():
         db.session.query(ContactQuestionnaire).delete()
         db.session.query(DemographicsQuestionnaire).delete()
         db.session.query(EvaluationHistoryQuestionnaire).delete()
-        db.session.query(GuardianDemographicsQuestionnaire).delete()
         db.session.query(HomeQuestionnaire).delete()
+        db.session.query(IdentificationQuestionnaire).delete()
         db.session.query(ResourceCategory).delete()
         db.session.query(StudyCategory).delete()
         db.session.query(TrainingCategory).delete()
