+++ backend/app/model/questionnaires/contact_questionnaire.py
@@ -12,66 +12,17 @@ class ContactQuestionnaire(db.Model):
     participant_id = db.Column(
         'participant_id',
         db.Integer,
+        db.ForeignKey('stardrive_participant.id')
+    )
+    user_id = db.Column(
+        'user_id',
+        db.Integer,
         db.ForeignKey('stardrive_user.id')
-    )
-
-    first_name = db.Column(
-        db.String,
-        info={
-            'display_order': 1,
-            'type': 'input',
-            'template_options': {
-                'label': 'First name',
-                'required': True
-            }
-        }
-    )
-
-    last_name = db.Column(
-        db.String,
-        info={
-            'display_order': 2,
-            'type': 'input',
-            'template_options': {
-                'label': 'Last name',
-                'required': True
-            }
-        }
-    )
-
-    is_first_name_preferred = db.Column(
-        db.Boolean,
-        info={
-            'display_order': 3,
-            'type': 'radio',
-            'default_value': True,
-            'template_options': {
-                'label': 'Is this your preferred name?',
-                'required': False,
-                'options': [
-                    {'value': True, 'label': 'Yes'},
-                    {'value': False, 'label': 'No'}
-                ]
-            }
-        }
-    )
-
-    nickname = db.Column(
-        db.String,
-        info={
-            'display_order': 4,
-            'type': 'input',
-            'template_options': {
-                'label': 'Nickname',
-                'required': False
-            },
-            'hide_expression': 'model.is_first_name_preferred'
-        }
     )
     phone = db.Column(
         db.String,
         info={
-            'display_order': 5.1,
+            'display_order': 1.1,
             'type': 'input',
             'template_options': {
                 'required': True,
@@ -83,11 +34,10 @@ class ContactQuestionnaire(db.Model):
             }
         }
     )
-
     phone_type = db.Column(
         db.String,
         info={
-            'display_order': 5.2,
+            'display_order': 1.2,
             'type': 'radio',
             'template_options': {
                 'label': '',
@@ -101,11 +51,10 @@ class ContactQuestionnaire(db.Model):
             }
         }
     )
-
     can_leave_voicemail = db.Column(
         db.Boolean,
         info={
-            'display_order': 5.3,
+            'display_order': 1.3,
             'type': 'radio',
             'default_value': True,
             'template_options': {
@@ -118,11 +67,10 @@ class ContactQuestionnaire(db.Model):
             }
         }
     )
-
     contact_times = db.Column(
         db.String,
         info={
-            'display_order': 5.4,
+            'display_order': 1.4,
             'type': 'textarea',
             'template_options': {
                 'label': 'Some research studies might involve a phone call. '
@@ -132,11 +80,10 @@ class ContactQuestionnaire(db.Model):
             },
         }
     )
-
     email = db.Column(
         db.String,
         info={
-            'display_order': 6,
+            'display_order': 2,
             'type': 'input',
             'template_options': {
                 'label': 'Email',
@@ -148,11 +95,10 @@ class ContactQuestionnaire(db.Model):
             },
         }
     )
-
     street_address = db.Column(
         db.String,
         info={
-            'display_order': 7.1,
+            'display_order': 3.1,
             'type': 'input',
             'template_options': {
                 'label': 'Street Address',
@@ -160,11 +106,10 @@ class ContactQuestionnaire(db.Model):
             }
         }
     )
-
     city = db.Column(
         db.String,
         info={
-            'display_order': 7.2,
+            'display_order': 3.2,
             'type': 'input',
             'template_options': {
                 'label': 'Town/City',
@@ -172,11 +117,10 @@ class ContactQuestionnaire(db.Model):
             }
         }
     )
-
     state = db.Column(
         db.String,
         info={
-            'display_order': 7.3,
+            'display_order': 3.3,
             'type': 'input',
             'template_options': {
                 'label': 'State',
@@ -184,11 +128,10 @@ class ContactQuestionnaire(db.Model):
             }
         }
     )
-
     zip = db.Column(
         db.Integer,
         info={
-            'display_order': 7.4,
+            'display_order': 3.4,
             'type': 'input',
             'template_options': {
                 'type': 'number',
@@ -200,11 +143,10 @@ class ContactQuestionnaire(db.Model):
             }
         }
     )
-
     marketing_channel = db.Column(
         db.String,
         info={
-            'display_order': 8.1,
+            'display_order': 4.1,
             'type': 'radio',
             'class_name': 'vertical-radio-group',
             'template_options': {
@@ -225,11 +167,10 @@ class ContactQuestionnaire(db.Model):
             }
         }
     )
-
     marketing_other = db.Column(
         db.String,
         info={
-            'display_order': 8.2,
+            'display_order': 4.2,
             'type': 'input',
             'template_options': {
                 'placeholder': 'Where did you hear about us?'
@@ -253,7 +194,7 @@ class ContactQuestionnaire(db.Model):
                         'can_leave_voicemail',
                         'contact_times'
                     ],
-                    'display_order': 5,
+                    'display_order': 1,
                     'wrappers': ['card'],
                     'template_options': {'label': 'Phone'},
                 },
@@ -264,7 +205,7 @@ class ContactQuestionnaire(db.Model):
                         'state',
                         'zip'
                     ],
-                    'display_order': 7,
+                    'display_order': 3,
                     'wrappers': ['card'],
                     'template_options': {'label': 'Address'},
                 },
@@ -273,7 +214,7 @@ class ContactQuestionnaire(db.Model):
                         'marketing_channel',
                         'marketing_other'
                     ],
-                    'display_order': 8,
+                    'display_order': 4,
                     'wrappers': ['card'],
                     'template_options': {'label': 'How did you hear about us?'},
                 }
@@ -290,8 +231,7 @@ class ContactQuestionnaire(db.Model):
 class ContactQuestionnaireSchema(ModelSchema):
     class Meta:
         model = ContactQuestionnaire
-        fields = ('id', 'last_updated', 'participant_id', 'first_name', 'last_name','is_first_name_preferred',
-                  'nickname', 'phone', 'phone_type', 'can_leave_voicemail', 'contact_times',
+        fields = ('id', 'last_updated', 'user_id', 'phone', 'phone_type', 'can_leave_voicemail', 'contact_times',
                   'email', 'street_address', 'city', 'state', 'zip', 'marketing_channel')
