+++ backend/app/resources/schema.py
@@ -24,8 +24,8 @@ import app.model.questionnaires.housemate
 import app.model.questionnaires.contact_questionnaire
 import app.model.questionnaires.demographics_questionnaire
 import app.model.questionnaires.evaluation_history_questionnaire
-import app.model.questionnaires.guardian_demographics_questionnaire
 import app.model.questionnaires.home_questionnaire
+import app.model.questionnaires.identification_questionnaire
 
 
 class OrganizationSchema(ModelSchema):
